Logic added to display lowest total energy at the end (which is the solution to the problem)
BurrowState.__init__ now assigns ID and lookup row for ID to self outside of the parameter specific logic

Added wrap option when writing BurrowState to log file.  If wrap is False the BurrowState will be shown in a single long line (which can be used in the search function in a text editor).  If wrap is True it multiple lines will be used to display the BurrowState (such as in the input given out by the developer of AdventOfCode).

Eliminated unnecessary repeat call to function energy_diff inside of verify_state_is_new

Added three new scenario input text files