Getting correct answer for part A.
Took steps intending to make BurrowState objects iterable.  Replaced all examples of list with tuple (except for immediate transitions when creating/updating the BurrowState).  Created BurrowState __hash__ function (note that __eq__ function was bundled in an earlier commit)

Eliminated all use of BurrowState's parent or children, since the packaged a-star algorithm now tracks which BurrowState is related to which other one.  Eliminated BurrowState's energy_total, because the a-star algorithm does not need that ... it will only use method distance_between().  Eliminated BurrowState's use of ID.

Counting total amphipods at the start.  When hashing a given BurrowState, it is assumed that every amphipod could be in one Sideroom.  Therefore this is used in function to hash siderooms.